# aspnetcore/src/IO.Swagger/Controllers/DefaultApi.cs
/*
 * Contact Emails
 *
 * Manages the emails used to contact Contacts.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: gielisg@selectsoftware.com.au
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DefaultApiController : ControllerBase
    { 
        /// <summary>
        /// Provides a list of Contact Email Mandatory Types Rules
        /// </summary>
        /// <remarks>List of Contact Email Mandatory Types. </remarks>
        /// <param name="sessionKey">The SessionKey for the current Session.</param>
        /// <param name="contactCode">Id. of the Contact</param>
        /// <response code="200">Successfully returned a list of a Contact Email Usage Mandatory Rules.</response>
        /// <response code="204">Resource Not Found, request was successfully processed but the requested resource was not found</response>
        /// <response code="400">Bad input parameter</response>
        /// <response code="401">Session expired</response>
        /// <response code="403">Authorisation Error</response>
        /// <response code="500">Internal server error</response>
        [HttpGet]
        [Route("/Selcomm/ContactEmails/1.0.0/ContactEmailMandatoryRules")]
        [ValidateModelState]
        [SwaggerOperation("ContactEmailMandatoryTypeList")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<ContactEmailMandatoryRule>), description: "Successfully returned a list of a Contact Email Usage Mandatory Rules.")]
        [SwaggerResponse(statusCode: 204, type: typeof(NotFound), description: "Resource Not Found, request was successfully processed but the requested resource was not found")]
        [SwaggerResponse(statusCode: 400, type: typeof(BadInputParameterError), description: "Bad input parameter")]
        [SwaggerResponse(statusCode: 401, type: typeof(SessionExpiredError), description: "Session expired")]
        [SwaggerResponse(statusCode: 403, type: typeof(AuthorisationError), description: "Authorisation Error")]
        [SwaggerResponse(statusCode: 500, type: typeof(InternalServerError), description: "Internal server error")]
        public virtual IActionResult ContactEmailMandatoryTypeList([FromHeader][Required()]string sessionKey, [FromQuery][Required()]string contactCode)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<ContactEmailMandatoryRule>));

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204, default(NotFound));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(BadInputParameterError));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(SessionExpiredError));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(AuthorisationError));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(InternalServerError));

            string exampleJson = null;
            exampleJson = "[ {\n  \"Type\" : \"Home Email\",\n  \"TypeCode\" : \"HE\"\n}, {\n  \"Type\" : \"Home Email\",\n  \"TypeCode\" : \"HE\"\n} ]";
            exampleJson = "<null>\n  <TypeCode>HE</TypeCode>\n  <Type>Home Email</Type>\n</null>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<ContactEmailMandatoryRule>>(exampleJson)
            : default(List<ContactEmailMandatoryRule>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Provides a list of Contact Email Usage items for a Contact
        /// </summary>
        /// <remarks>List of Contact Email Usage. </remarks>
        /// <param name="sessionKey">The SessionKey for the current Session.</param>
        /// <param name="contactCode">Id. of the Contact</param>
        /// <param name="includeTypes">TRUE will include list of available Email Type Codes, FALSE will not.</param>
        /// <param name="includeMandatoryRules">TRUE will include list of Mandatory Email Type Codes, FALSE will not.</param>
        /// <response code="200">Contact Email Usage</response>
        /// <response code="204">Resource Not Found, request was successfully processed but the requested resource was not found</response>
        /// <response code="400">Bad input parameter</response>
        /// <response code="401">Session expired</response>
        /// <response code="403">Authorisation Error</response>
        /// <response code="500">Internal server error</response>
        [HttpGet]
        [Route("/Selcomm/ContactEmails/1.0.0/ContactEmailUsage")]
        [ValidateModelState]
        [SwaggerOperation("ContactEmailUsage")]
        [SwaggerResponse(statusCode: 200, type: typeof(ContactEmailUsage), description: "Contact Email Usage")]
        [SwaggerResponse(statusCode: 204, type: typeof(NotFound), description: "Resource Not Found, request was successfully processed but the requested resource was not found")]
        [SwaggerResponse(statusCode: 400, type: typeof(BadInputParameterError), description: "Bad input parameter")]
        [SwaggerResponse(statusCode: 401, type: typeof(SessionExpiredError), description: "Session expired")]
        [SwaggerResponse(statusCode: 403, type: typeof(AuthorisationError), description: "Authorisation Error")]
        [SwaggerResponse(statusCode: 500, type: typeof(InternalServerError), description: "Internal server error")]
        public virtual IActionResult ContactEmailUsage([FromHeader][Required()]string sessionKey, [FromQuery][Required()]string contactCode, [FromQuery][Required()]bool? includeTypes, [FromQuery][Required()]bool? includeMandatoryRules)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ContactEmailUsage));

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204, default(NotFound));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(BadInputParameterError));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(SessionExpiredError));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(AuthorisationError));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(InternalServerError));

            string exampleJson = null;
            exampleJson = "{\n  \"ContactEmailUsage\" : [ {\n    \"UpdatedBy\" : 4001304,\n    \"Id\" : 1,\n    \"LastUpdated\" : \"2019-02-01 00:00:00\",\n    \"EmailAddress\" : \"gielisg@selectsoftware.com.au\",\n    \"FromDateTime\" : \"2019-02-01 00:00:00\",\n    \"EmailTypes\" : [ {\n      \"Code\" : \"HP\",\n      \"Name\" : \"Home Email\"\n    }, {\n      \"Code\" : \"HP\",\n      \"Name\" : \"Home Email\"\n    } ]\n  }, {\n    \"UpdatedBy\" : 4001304,\n    \"Id\" : 1,\n    \"LastUpdated\" : \"2019-02-01 00:00:00\",\n    \"EmailAddress\" : \"gielisg@selectsoftware.com.au\",\n    \"FromDateTime\" : \"2019-02-01 00:00:00\",\n    \"EmailTypes\" : [ {\n      \"Code\" : \"HP\",\n      \"Name\" : \"Home Email\"\n    }, {\n      \"Code\" : \"HP\",\n      \"Name\" : \"Home Email\"\n    } ]\n  } ],\n  \"ContactEmailTypes\" : [ {\n    \"Code\" : \"HP\",\n    \"Name\" : \"Home Email\"\n  }, {\n    \"Code\" : \"HP\",\n    \"Name\" : \"Home Email\"\n  } ],\n  \"ContactEmailMandatoryRules\" : [ {\n    \"Type\" : \"Home Email\",\n    \"TypeCode\" : \"HE\"\n  }, {\n    \"Type\" : \"Home Email\",\n    \"TypeCode\" : \"HE\"\n  } ]\n}";
            exampleJson = "<null>\n</null>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<ContactEmailUsage>(exampleJson)
            : default(ContactEmailUsage);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Provides history of Contact Email Usage items for a Contact
        /// </summary>
        /// <remarks>List of Contact Email Usage. </remarks>
        /// <param name="sessionKey">The SessionKey for the current Session.</param>
        /// <param name="contactCode">Id. of the Contact</param>
        /// <response code="200">Contact Email Usage History</response>
        /// <response code="204">Resource Not Found, request was successfully processed but the requested resource was not found</response>
        /// <response code="400">Bad input parameter</response>
        /// <response code="401">Session expired</response>
        /// <response code="403">Authorisation Error</response>
        /// <response code="500">Internal server error</response>
        [HttpGet]
        [Route("/Selcomm/ContactEmails/1.0.0/ContactEmailUsageHistory")]
        [ValidateModelState]
        [SwaggerOperation("ContactEmailUsageHistory")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<ContactEmailHistory>), description: "Contact Email Usage History")]
        [SwaggerResponse(statusCode: 204, type: typeof(NotFound), description: "Resource Not Found, request was successfully processed but the requested resource was not found")]
        [SwaggerResponse(statusCode: 400, type: typeof(BadInputParameterError), description: "Bad input parameter")]
        [SwaggerResponse(statusCode: 401, type: typeof(SessionExpiredError), description: "Session expired")]
        [SwaggerResponse(statusCode: 403, type: typeof(AuthorisationError), description: "Authorisation Error")]
        [SwaggerResponse(statusCode: 500, type: typeof(InternalServerError), description: "Internal server error")]
        public virtual IActionResult ContactEmailUsageHistory([FromHeader][Required()]string sessionKey, [FromQuery][Required()]string contactCode)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<ContactEmailHistory>));

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204, default(NotFound));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(BadInputParameterError));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(SessionExpiredError));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(AuthorisationError));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(InternalServerError));

            string exampleJson = null;
            exampleJson = "[ {\n  \"UpdatedBy\" : 4001304,\n  \"EmailType\" : \"Mobile\",\n  \"Id\" : 1,\n  \"ToDateTime\" : \"2019-04-01 00:00:00\",\n  \"EmailTypeCode\" : \"HP\",\n  \"EmailAddress\" : \"gielisg@selectsoftware.com.au\",\n  \"FromDateTime\" : \"2019-02-01 00:00:00\"\n}, {\n  \"UpdatedBy\" : 4001304,\n  \"EmailType\" : \"Mobile\",\n  \"Id\" : 1,\n  \"ToDateTime\" : \"2019-04-01 00:00:00\",\n  \"EmailTypeCode\" : \"HP\",\n  \"EmailAddress\" : \"gielisg@selectsoftware.com.au\",\n  \"FromDateTime\" : \"2019-02-01 00:00:00\"\n} ]";
            exampleJson = "<null>\n  <Id>1</Id>\n  <EmailTypeCode>HP</EmailTypeCode>\n  <EmailType>Mobile</EmailType>\n  <EmailAddress>gielisg@selectsoftware.com.au</EmailAddress>\n</null>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<ContactEmailHistory>>(exampleJson)
            : default(List<ContactEmailHistory>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Updates Email Usage for a Contact
        /// </summary>
        /// <remarks>Updates Contact Email Usage. </remarks>
        /// <param name="sessionKey">The SessionKey for the current Session.</param>
        /// <param name="contactEmailUsageUpdate">Contact Email Usage Update</param>
        /// <response code="200">ContactEmailUsageUpdateResponse</response>
        /// <response code="204">Resource Not Found, request was successfully processed but the requested resource was not found</response>
        /// <response code="400">Bad input parameter</response>
        /// <response code="401">Session expired</response>
        /// <response code="403">Authorisation Error</response>
        /// <response code="500">Internal server error</response>
        [HttpPost]
        [Route("/Selcomm/ContactEmails/1.0.0/ContactEmailUsageUpdate")]
        [ValidateModelState]
        [SwaggerOperation("ContactEmailUsageUpdate")]
        [SwaggerResponse(statusCode: 200, type: typeof(InlineResponse200), description: "ContactEmailUsageUpdateResponse")]
        [SwaggerResponse(statusCode: 204, type: typeof(NotFound), description: "Resource Not Found, request was successfully processed but the requested resource was not found")]
        [SwaggerResponse(statusCode: 400, type: typeof(BadInputParameterError), description: "Bad input parameter")]
        [SwaggerResponse(statusCode: 401, type: typeof(SessionExpiredError), description: "Session expired")]
        [SwaggerResponse(statusCode: 403, type: typeof(AuthorisationError), description: "Authorisation Error")]
        [SwaggerResponse(statusCode: 500, type: typeof(InternalServerError), description: "Internal server error")]
        public virtual IActionResult ContactEmailUsageUpdate([FromHeader][Required()]string sessionKey, [FromBody]ContactEmailUsageUpdate contactEmailUsageUpdate)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(InlineResponse200));

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204, default(NotFound));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(BadInputParameterError));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(SessionExpiredError));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(AuthorisationError));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(InternalServerError));

            string exampleJson = null;
            exampleJson = "{\n  \"ErrorMessage\" : \"Missing mandatory Email Type\",\n  \"Successful\" : true\n}";
            exampleJson = "<null>\n  <Successful>true</Successful>\n  <ErrorMessage>Missing mandatory Email Type</ErrorMessage>\n</null>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<InlineResponse200>(exampleJson)
            : default(InlineResponse200);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Provides a list of Contact Phone Types
        /// </summary>
        /// <remarks>List of Contact Email Types to support GUI selection List </remarks>
        /// <param name="sessionKey">The SessionKey for the current Session.</param>
        /// <response code="200">Successfully returned a list of a Contact Email Types.</response>
        /// <response code="204">Resource Not Found, request was successfully processed but the requested resource was not found</response>
        /// <response code="400">Bad input parameter</response>
        /// <response code="401">Session expired</response>
        /// <response code="403">Authorisation Error</response>
        /// <response code="500">Internal server error</response>
        [HttpGet]
        [Route("/Selcomm/ContactEmails/1.0.0/ContactEmailTypes")]
        [ValidateModelState]
        [SwaggerOperation("ContactPhoneTypeList")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<ContactEmailType>), description: "Successfully returned a list of a Contact Email Types.")]
        [SwaggerResponse(statusCode: 204, type: typeof(NotFound), description: "Resource Not Found, request was successfully processed but the requested resource was not found")]
        [SwaggerResponse(statusCode: 400, type: typeof(BadInputParameterError), description: "Bad input parameter")]
        [SwaggerResponse(statusCode: 401, type: typeof(SessionExpiredError), description: "Session expired")]
        [SwaggerResponse(statusCode: 403, type: typeof(AuthorisationError), description: "Authorisation Error")]
        [SwaggerResponse(statusCode: 500, type: typeof(InternalServerError), description: "Internal server error")]
        public virtual IActionResult ContactPhoneTypeList([FromHeader][Required()]string sessionKey)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<ContactEmailType>));

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204, default(NotFound));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(BadInputParameterError));

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401, default(SessionExpiredError));

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403, default(AuthorisationError));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(InternalServerError));

            string exampleJson = null;
            exampleJson = "[ {\n  \"Code\" : \"HP\",\n  \"Name\" : \"Home Email\"\n}, {\n  \"Code\" : \"HP\",\n  \"Name\" : \"Home Email\"\n} ]";
            exampleJson = "<null>\n  <Code>HP</Code>\n  <Name>Home Email</Name>\n</null>";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<ContactEmailType>>(exampleJson)
            : default(List<ContactEmailType>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}