# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Manages the emails used to contact Contacts.",
    "version" : "1.0.0",
    "title" : "Contact Emails",
    "contact" : {
      "email" : "gielisg@selectsoftware.com.au"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/Selcomm/ContactEmails/1.0.0",
  "schemes" : [ "https" ],
  "paths" : {
    "/ContactEmailTypes" : {
      "get" : {
        "summary" : "Provides a list of Contact Phone Types",
        "description" : "List of Contact Email Types to support GUI selection List\n",
        "operationId" : "ContactPhoneTypeList",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "SessionKey",
          "in" : "header",
          "description" : "The SessionKey for the current Session.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully returned a list of a Contact Email Types.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ContactEmailType"
              }
            }
          },
          "204" : {
            "description" : "Resource Not Found, request was successfully processed but the requested resource was not found",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "400" : {
            "description" : "Bad input parameter",
            "schema" : {
              "$ref" : "#/definitions/BadInputParameterError"
            }
          },
          "401" : {
            "description" : "Session expired",
            "schema" : {
              "$ref" : "#/definitions/SessionExpiredError"
            }
          },
          "403" : {
            "description" : "Authorisation Error",
            "schema" : {
              "$ref" : "#/definitions/AuthorisationError"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/ContactEmailMandatoryRules" : {
      "get" : {
        "summary" : "Provides a list of Contact Email Mandatory Types Rules",
        "description" : "List of Contact Email Mandatory Types.\n",
        "operationId" : "ContactEmailMandatoryTypeList",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "SessionKey",
          "in" : "header",
          "description" : "The SessionKey for the current Session.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "ContactCode",
          "in" : "query",
          "description" : "Id. of the Contact",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully returned a list of a Contact Email Usage Mandatory Rules.",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ContactEmailMandatoryRule"
              }
            }
          },
          "204" : {
            "description" : "Resource Not Found, request was successfully processed but the requested resource was not found",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "400" : {
            "description" : "Bad input parameter",
            "schema" : {
              "$ref" : "#/definitions/BadInputParameterError"
            }
          },
          "401" : {
            "description" : "Session expired",
            "schema" : {
              "$ref" : "#/definitions/SessionExpiredError"
            }
          },
          "403" : {
            "description" : "Authorisation Error",
            "schema" : {
              "$ref" : "#/definitions/AuthorisationError"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/ContactEmailUsage" : {
      "get" : {
        "summary" : "Provides a list of Contact Email Usage items for a Contact",
        "description" : "List of Contact Email Usage.\n",
        "operationId" : "ContactEmailUsage",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "SessionKey",
          "in" : "header",
          "description" : "The SessionKey for the current Session.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "ContactCode",
          "in" : "query",
          "description" : "Id. of the Contact",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "IncludeTypes",
          "in" : "query",
          "description" : "TRUE will include list of available Email Type Codes, FALSE will not.",
          "required" : true,
          "type" : "boolean",
          "default" : false
        }, {
          "name" : "IncludeMandatoryRules",
          "in" : "query",
          "description" : "TRUE will include list of Mandatory Email Type Codes, FALSE will not.",
          "required" : true,
          "type" : "boolean",
          "default" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "Contact Email Usage",
            "schema" : {
              "$ref" : "#/definitions/ContactEmailUsage"
            }
          },
          "204" : {
            "description" : "Resource Not Found, request was successfully processed but the requested resource was not found",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "400" : {
            "description" : "Bad input parameter",
            "schema" : {
              "$ref" : "#/definitions/BadInputParameterError"
            }
          },
          "401" : {
            "description" : "Session expired",
            "schema" : {
              "$ref" : "#/definitions/SessionExpiredError"
            }
          },
          "403" : {
            "description" : "Authorisation Error",
            "schema" : {
              "$ref" : "#/definitions/AuthorisationError"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/ContactEmailUsageUpdate" : {
      "post" : {
        "summary" : "Updates Email Usage for a Contact",
        "description" : "Updates Contact Email Usage.\n",
        "operationId" : "ContactEmailUsageUpdate",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "SessionKey",
          "in" : "header",
          "description" : "The SessionKey for the current Session.",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "ContactEmailUsageUpdate",
          "description" : "Contact Email Usage Update",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ContactEmailUsageUpdate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "ContactEmailUsageUpdateResponse",
            "schema" : {
              "$ref" : "#/definitions/inline_response_200"
            }
          },
          "204" : {
            "description" : "Resource Not Found, request was successfully processed but the requested resource was not found",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "400" : {
            "description" : "Bad input parameter",
            "schema" : {
              "$ref" : "#/definitions/BadInputParameterError"
            }
          },
          "401" : {
            "description" : "Session expired",
            "schema" : {
              "$ref" : "#/definitions/SessionExpiredError"
            }
          },
          "403" : {
            "description" : "Authorisation Error",
            "schema" : {
              "$ref" : "#/definitions/AuthorisationError"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/ContactEmailUsageHistory" : {
      "get" : {
        "summary" : "Provides history of Contact Email Usage items for a Contact",
        "description" : "List of Contact Email Usage.\n",
        "operationId" : "ContactEmailUsageHistory",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "SessionKey",
          "in" : "header",
          "description" : "The SessionKey for the current Session.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "ContactCode",
          "in" : "query",
          "description" : "Id. of the Contact",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Contact Email Usage History",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ContactEmailHistory"
              }
            }
          },
          "204" : {
            "description" : "Resource Not Found, request was successfully processed but the requested resource was not found",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "400" : {
            "description" : "Bad input parameter",
            "schema" : {
              "$ref" : "#/definitions/BadInputParameterError"
            }
          },
          "401" : {
            "description" : "Session expired",
            "schema" : {
              "$ref" : "#/definitions/SessionExpiredError"
            }
          },
          "403" : {
            "description" : "Authorisation Error",
            "schema" : {
              "$ref" : "#/definitions/AuthorisationError"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "ContactEmailType" : {
      "properties" : {
        "Code" : {
          "type" : "string",
          "example" : "HP",
          "maxLength" : 2
        },
        "Name" : {
          "type" : "string",
          "example" : "Home Email",
          "description" : "Name of the Contact Email Type."
        }
      },
      "example" : {
        "Code" : "HP",
        "Name" : "Home Email"
      }
    },
...
    "NotFound" : {
      "type" : "object",
      "properties" : {
        "ErrorCode" : {
          "type" : "number",
          "example" : 204.0,
          "description" : "The status code, same as http status"
        },
        "ErrorMessage" : {
          "type" : "string",
          "example" : "Resource not found"
        },
        "DetailedInformation" : {
          "type" : "string",
          "example" : "Database entity not found",
          "description" : "Additional details"
        }
      },
      "description" : "Response object for a Resource Not Found condition"
    },
    "BadInputParameterError" : {
      "type" : "object",
      "properties" : {
        "ErrorCode" : {
          "type" : "number",
          "example" : 400.0,
          "description" : "The error code, same as http status"
        },
        "ErrorMessage" : {
          "type" : "string",
          "example" : "Bad Input Parameter"
        },
        "DetailedInformation" : {
          "type" : "string",
          "example" : "Quantity parameter should be numeric",
          "description" : "Additional error details"
        }
      },
      "description" : "Error object for bad input parameters"
    },
    "SessionExpiredError" : {
      "type" : "object",
      "properties" : {
        "ErrorCode" : {
          "type" : "number",
          "example" : 401.0,
          "description" : "The error code, same as http status"
        },
        "ErrorMessage" : {
          "type" : "string",
          "example" : "Session expired"
        },
        "DetailedInformation" : {
          "type" : "string",
          "example" : "Session has expired, please re-authenticate",
          "description" : "Additional error details"
        }
      },
      "description" : "Error object for a Session Expired Error"
    },
    "AuthorisationError" : {
      "type" : "object",
      "properties" : {
        "ErrorCode" : {
          "type" : "number",
          "example" : 403.0,
          "description" : "The error code, same as http status"
        },
        "ErrorMessage" : {
          "type" : "string",
          "example" : "Resource Denied"
        },
        "DetailedInformation" : {
          "type" : "string",
          "example" : "User not authorised to use the ChangePassword API",
          "description" : "Additional error details"
        }
      },
      "description" : "Error object for a Authorisation Error"
    },
    "InternalServerError" : {
      "type" : "object",
      "properties" : {
        "ErrorCode" : {
          "type" : "number",
          "example" : 500.0,
          "description" : "The error code, same as http status"
        },
        "ErrorMessage" : {
          "type" : "string",
          "example" : "Internal Server Error"
        },
        "DetailedInformation" : {
          "type" : "string",
          "example" : "Database refusing connection requests",
          "description" : "Additional error details"
        }
      },
      "description" : "Error object for an internal server error"
    },
...
  }
}